Refactored to use dynamic home page UI
