+++ src/middleware.ts
@@ -1,6 +1,6 @@
 import { clerkMiddleware, createRouteMatcher } from '@clerk/nextjs/server'
 
-const isProtectedRoute = createRouteMatcher(['/chat', '/chat(.*)'])
+const isProtectedRoute = createRouteMatcher([])
 
 export default clerkMiddleware(
 	(auth, req) => {
